+++ NEWS.rst
@@ -6,6 +6,8 @@ of every change, see the Git log.
 
 Latest
 ------
+* Major: Remove the explict dependency on virtualenv. Instead we will
+  automatically download the virtualenv dependency.
 * Minor: Updated waf to 2.0.10.
 * Patch: Hide the output of the mklink command on Windows when calling the
   create_symlink function (the output is logged when an error occurs)